requirements.txt
Flask
Flask-SQLAlchemy
Authlib
Pillow
gunicorn
psycopg2-binary